+++ src/manifolds/SymmetricPositiveDefinite.jl
@@ -4,9 +4,8 @@
 The manifold of symmetric positive definite matrices, i.e.
 
 ````math
-\mathcal P(n) =
-\bigl\{
-p ∈ ℝ^{n×n}\ \big|\ a^\mathrm{T}pa > 0 \text{ for all } a ∈ ℝ^{n}\backslash\{0\}
+\mathcal P(n) = \bigl\{
+  p ∈ ℝ^{n×n}\ \big|\ a^\mathrm{T}pa > 0 \text{ for all } a ∈ ℝ^{n}\backslash\{0\}
 \bigr\}
 ````
 
@@ -14,9 +13,7 @@ The tangent space at ``T_p\mathcal P(n)`` reads
 
 ```math
     T_p\mathcal P(n) =
-    \bigl\{
-        X \in \mathbb R^{n×n} \big|\ X=X^\mathrm{T}
-    \bigr\},
+    \bigl\{ X \in \mathbb R^{n×n} \big|\ X=X^\mathrm{T} \bigr\},
 ```
 i.e. the set of symmetric matrices,
 
@@ -26,9 +23,7 @@ i.e. the set of symmetric matrices,
 
 generates the manifold ``\mathcal P(n) \subset ℝ^{n×n}``
 """
-struct SymmetricPositiveDefinite{T} <: AbstractDecoratorManifold{ℝ}
-    size::T
-end
+const SymmetricPositiveDefinite{T} = HermitianPositiveDefinite{ℝ, T}
 
 function SymmetricPositiveDefinite(n::Int; parameter::Symbol = :type)
     size = wrap_type_parameter(parameter, (n,))